Move otconfig parser unit tests to own test module
