+++ tests/OTAnalytics/plugin_parser/test_otvision_parser.py
@@ -1,15 +1,12 @@
 from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, Sequence
 from unittest.mock import Mock, call
 
 import pytest
 
 from OTAnalytics import version
-from OTAnalytics.application.datastore import FlowParser, VideoMetadata, VideoParser
-from OTAnalytics.application.parser.config_parser import OtConfig
-from OTAnalytics.application.project import Project
-from OTAnalytics.domain import flow, geometry, section, video
+from OTAnalytics.application.datastore import VideoMetadata, VideoParser
+from OTAnalytics.domain import flow, geometry, section
 from OTAnalytics.domain.event import EVENT_LIST, Event, EventType
 from OTAnalytics.domain.flow import Flow, FlowId
 from OTAnalytics.domain.geometry import (
@@ -40,8 +37,7 @@ from OTAnalytics.plugin_datastore.python_track_store import (
     PythonTrackDataset,
 )
 from OTAnalytics.plugin_parser import dataformat_versions, ottrk_dataformat
-from OTAnalytics.plugin_parser.json_parser import parse_json, write_json, write_json_bz2
-from OTAnalytics.plugin_parser.otconfig_parser import PROJECT, OtConfigParser
+from OTAnalytics.plugin_parser.json_parser import write_json, write_json_bz2
 from OTAnalytics.plugin_parser.otvision_parser import (
     DEFAULT_TRACK_LENGTH_LIMIT,
     EVENT_FORMAT_VERSION,
@@ -767,82 +763,3 @@ class TestCachedVideoParser:
         result = cached_parser.convert([video1, video2], test_data_tmp_dir)
 
         assert expected_result is result
-
-
-class TestOtConfigParser:
-    def test_serialize_config(self, test_data_tmp_dir: Path) -> None:
-        video_parser = Mock(spec=VideoParser)
-        flow_parser = Mock(spec=FlowParser)
-        config_parser = OtConfigParser(
-            video_parser=video_parser,
-            flow_parser=flow_parser,
-        )
-        project = Project(name="My Test Project", start_date=datetime(2020, 1, 1))
-        videos: list[Video] = []
-        sections: list[Section] = []
-        flows: list[Flow] = []
-        output = test_data_tmp_dir / "config.otconfig"
-        serialized_videos = {video.VIDEOS: {"serialized": "videos"}}
-        serialized_sections = {section.SECTIONS: {"serialized": "sections"}}
-        video_parser.convert.return_value = serialized_videos
-        flow_parser.convert.return_value = serialized_sections
-
-        config_parser.serialize(
-            project=project,
-            video_files=videos,
-            sections=sections,
-            flows=flows,
-            file=output,
-        )
-
-        serialized_content = parse_json(output)
-        expected_content: dict[str, Any] = {PROJECT: project.to_dict()}
-        expected_content |= serialized_videos
-        expected_content |= serialized_sections
-
-        assert serialized_content == expected_content
-        assert video_parser.convert.call_args_list == [
-            call(videos, relative_to=test_data_tmp_dir)
-        ]
-        assert flow_parser.convert.call_args_list == [call(sections, flows)]
-
-    def test_parse_config(self, test_data_tmp_dir: Path) -> None:
-        video_parser = Mock(spec=VideoParser)
-        flow_parser = Mock(spec=FlowParser)
-        config_parser = OtConfigParser(
-            video_parser=video_parser,
-            flow_parser=flow_parser,
-        )
-        project = Project(
-            name="Test Project", start_date=datetime(2020, 1, 1, tzinfo=timezone.utc)
-        )
-        videos: Sequence[Video] = ()
-        sections: Sequence[Section] = ()
-        flows: Sequence[Flow] = ()
-        config_file = test_data_tmp_dir / "config.otconfig"
-        serialized_videos = {video.VIDEOS: {"serialized": "videos"}}
-        serialized_flows = {
-            section.SECTIONS: {"serialized": "sections"},
-            flow.FLOWS: {"serialized": "flows"},
-        }
-        video_parser.convert.return_value = serialized_videos
-        flow_parser.convert.return_value = serialized_flows
-        video_parser.parse_list.return_value = videos
-        flow_parser.parse_content.return_value = sections, flows
-
-        config_parser.serialize(
-            project=project,
-            video_files=videos,
-            sections=sections,
-            flows=flows,
-            file=config_file,
-        )
-        config = config_parser.parse(file=config_file)
-
-        expected_config = OtConfig(
-            project=project,
-            videos=videos,
-            sections=sections,
-            flows=flows,
-        )
-        assert config == expected_config
